check application permissions only if create a project and not in each scan creation
