+++ internal/services/projects.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	errorConstants "github.com/checkmarx/ast-cli/internal/constants/errors"
 	"slices"
 	"strconv"
 	"time"
@@ -22,7 +23,6 @@ const (
 )
 
 func FindProject(
-	applicationID []string,
 	projectName string,
 	cmd *cobra.Command,
 	projectsWrapper wrappers.ProjectsWrapper,
@@ -54,6 +54,13 @@ func FindProject(
 
 	projectGroups, _ := cmd.Flags().GetString(commonParams.ProjectGroupList)
 	projectPrivatePackage, _ := cmd.Flags().GetString(commonParams.ProjecPrivatePackageFlag)
+
+	applicationName, _ := cmd.Flags().GetString(commonParams.ApplicationName)
+	applicationID, appErr := getApplicationId(applicationName, applicationWrapper)
+	if appErr != nil {
+		return "", appErr
+	}
+
 	projectID, err := createProject(projectName, cmd, projectsWrapper, groupsWrapper, accessManagementWrapper, applicationWrapper,
 		applicationID, projectGroups, projectPrivatePackage, featureFlagsWrapper)
 	if err != nil {
@@ -84,6 +91,49 @@ func GetProjectsCollectionByProjectName(projectName string, projectsWrapper wrap
 	}
 
 	return resp, nil
+}
+
+func getApplicationId(applicationName string, applicationsWrapper wrappers.ApplicationsWrapper) ([]string, error) {
+	var applicationID []string
+	if applicationName != "" {
+		application, getAppErr := getApplication(applicationName, applicationsWrapper)
+		if getAppErr != nil {
+			return nil, getAppErr
+		}
+		if application == nil {
+			return nil, errors.Errorf(errorConstants.ApplicationDoesntExistOrNoPermission)
+		}
+		applicationID = []string{application.ID}
+	}
+	return applicationID, nil
+}
+
+func getApplication(applicationName string, applicationsWrapper wrappers.ApplicationsWrapper) (*wrappers.Application, error) {
+	if applicationName != "" {
+		params := make(map[string]string)
+		params["name"] = applicationName
+		resp, err := applicationsWrapper.Get(params)
+		if err != nil {
+			return nil, err
+		}
+		if resp.Applications != nil && len(resp.Applications) > 0 {
+			application := verifyApplicationNameExactMatch(applicationName, resp)
+
+			return application, nil
+		}
+	}
+	return nil, nil
+}
+
+func verifyApplicationNameExactMatch(applicationName string, resp *wrappers.ApplicationsResponseModel) *wrappers.Application {
+	var application *wrappers.Application
+	for i := range resp.Applications {
+		if resp.Applications[i].Name == applicationName {
+			application = &resp.Applications[i]
+			break
+		}
+	}
+	return application
 }
 
 func createProject(
